+++ Source/AIFactoryCopilot/Public/AIFactoryCreativeNodePlacement.h
@@ -2,6 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "Resources/FGResourceNode.h"
+#include "Resources/FGResourceNodeBase.h"
 
 class AFGPlayerController;
 class UFGResourceDescriptor;
@@ -13,5 +14,12 @@ namespace AIFactoryCreativeNodePlacement
         AFGPlayerController* PlayerController,
         TSubclassOf<UFGResourceDescriptor> Resource,
         EResourcePurity Purity,
+        FString& OutReason);
+
+    bool ArmForPlayer(
+        AFGPlayerController* PlayerController,
+        TSubclassOf<UFGResourceDescriptor> Resource,
+        EResourcePurity Purity,
+        EResourceNodeType NodeType,
         FString& OutReason);
 }